Pass driver config as generic.

+++ packages/driver-typescript/src/TypeScriptDriver.ts
@@ -6,10 +6,11 @@
 import path from 'path';
 import rimraf from 'rimraf';
 import { Argv, Driver, DriverContext } from '@beemo/core';
+import { TypeScriptConfig } from './types';
 
 // Success: Writes nothing to stdout or stderr
 // Failure: Writes to stdout on syntax and type error
-export default class TypeScriptDriver extends Driver {
+export default class TypeScriptDriver extends Driver<TypeScriptConfig> {
   bootstrap() {
     this.setMetadata({
       bin: 'tsc',
@@ -35,8 +36,12 @@ export default class TypeScriptDriver extends Driver {
   /**
    * Automatically clean the target folder if `outDir` is used.
    */
-  handleCleanTarget = (driver: Driver, argv: Argv, { args }: DriverContext) => {
-    const { outDir } = this.options;
+  handleCleanTarget = (driver: TypeScriptDriver, argv: Argv, { args }: DriverContext) => {
+    if (!this.config.compilerOptions) {
+      return;
+    }
+
+    const { outDir } = this.config.compilerOptions;
 
     if (args.clean && outDir) {
       rimraf.sync(path.resolve(outDir));
